sui_replay::batch_replay
Function batch_replay

pub async fn batch_replay(
    tx_digests: impl Iterator<Item = TransactionDigest>,
    num_tasks: u64,
    rpc_url: String,
    expensive_safety_check_config: ExpensiveSafetyCheckConfig,
    use_authority: bool,
    terminate_early: bool,
    persist_path: Option<PathBuf>,
)

Given a list of transaction digests, replay them in parallel using num_tasks tasks. If terminate_early is true, the replay will terminate early if any transaction fails; otherwise it will try to finish all transactions.